extract account manager
