+++ management/server/token_mgr.go
@@ -9,9 +9,9 @@ import (
 	"sync"
 	"time"
 
+	"github.com/netbirdio/management-integrations/integrations"
 	log "github.com/sirupsen/logrus"
 
-	"github.com/netbirdio/management-integrations/integrations/config"
 	"github.com/netbirdio/netbird/management/proto"
 	auth "github.com/netbirdio/netbird/relay/auth/hmac"
 	authv2 "github.com/netbirdio/netbird/relay/auth/hmac/v2"
@@ -217,7 +217,7 @@ func (m *TimeBasedAuthSecretsManager) pushNewTURNAndRelayTokens(ctx context.Cont
 		}
 	}
 
-	config.ExtendNetBirdConfig(update.NetbirdConfig, nil)
+	integrations.ExtendNetBirdConfig(update.NetbirdConfig, nil)
 
 	log.WithContext(ctx).Debugf("sending new TURN credentials to peer %s", peerID)
 	m.updateManager.SendUpdate(ctx, peerID, &UpdateMessage{Update: update})
@@ -241,7 +241,7 @@ func (m *TimeBasedAuthSecretsManager) pushNewRelayTokens(ctx context.Context, pe
 		},
 	}
 
-	config.ExtendNetBirdConfig(update.NetbirdConfig, nil)
+	integrations.ExtendNetBirdConfig(update.NetbirdConfig, nil)
 
 	log.WithContext(ctx).Debugf("sending new relay credentials to peer %s", peerID)
 	m.updateManager.SendUpdate(ctx, peerID, &UpdateMessage{Update: update})
